Some example code to look at
As an example I adapted a bash script into python over three
iterations. In the first attempt I just did a quick and dirty
attempt to get the core functionality working. In the second
iteration things improve with error checking and better printing
and by the third attempt, it starts to look like a real program.