Feedback on the cleaner script
Your script works great, but I modified the input section to
be more in keeping with how Linux tools normally work. Generally
if you need input like a directory name, string to modify, etc
you would just pass it in on the commandline instead of specifically
asking the user for it. You want to prompt the user in situations
that would require extensive user input. For instance, if they were
filling out a form, putting their life's information on the
command line would be impractical. However, for most of our
purposes, passing in an argument will be more than sufficient.