Tweaking the tweet checker
So the basic idea here is solid. The code works and does what
we need it to do. So this commit adds two things. First, it
changes the print statements slightly to make it easier to
read and type. We don't need a bunch of plus signs everywhere
because as things get larger we don't want to have to type
this all the time.

Next, I added in a little bit of extra output so we can see
how many characters we are checking on each line. This is
just a trouble shooting tip so we can see if any lines have
0 chars (which would be bad) or if something weird is
happening.